refactor: magicNumber, magicString 생성 및 적용
