+++ frontend/src/api/get.ts
@@ -1,5 +1,6 @@
 import { api, customerHeader, ownerHeader } from '.';
-import { PARAMS_ERROR_MESSAGE } from '../constants';
+import { PARAMS_ERROR_MESSAGE } from '../constants/magicString';
+
 import {
   QueryReq,
   PhoneNumberParams,
