+++ frontend/src/constants/magicNumber.ts
@@ -1,0 +1,19 @@
+export const PHONE_NUMBER_LENGTH = 13;
+
+export const ENTER_KEY_INDEX = 11;
+
+export const BACK_KEY_INDEX = 9;
+
+export const EXPIRE_DATE_MAX = 1200;
+
+export const INVALID_CAFE_ID = -1;
+
+export const MEGA_BYTE = 1024 ** 2;
+
+export const IMAGE_MAX_SIZE = 5 * MEGA_BYTE;
+
+export const INTRO_LIMITATION = 150;
+
+export const PHONE_NUMBER_START_WITH_02 = 12;
+
+export const PHONE_NUMBER_START_WITHOUT_02 = 13;
