+++ frontend/src/constants/magicString.ts
@@ -1,0 +1,3 @@
+export const EXPIRE_DATE_NONE = '없음';
+
+export const PARAMS_ERROR_MESSAGE = '[ERROR] params를 지정해주세요.';
