+++ frontend/src/pages/Admin/CouponDesign/ChoiceTemplate/index.tsx
@@ -1,13 +1,19 @@
 import { Dispatch, SetStateAction, useState } from 'react';
 import { ChoiceTemplateContainer } from './style';
 import TabBar from '../../../../components/TabBar';
-import { TEMPLATE_MENU } from '../../../../constants';
 import {
   SampleBackCouponImage,
   SampleImage,
   StampCoordinate,
 } from '../../../../types/domain/coupon';
 import SampleImageList from './SampleImageList';
+import { TemplateMenu } from '../../../../types/utils';
+
+export const TEMPLATE_MENU: Record<string, TemplateMenu> = {
+  FRONT_IMAGE: '쿠폰(앞)',
+  BACK_IMAGE: '쿠폰(뒤)',
+  STAMP: '스탬프',
+};
 
 export const TEMPLATE_OPTIONS = [
   {
