+++ frontend/src/pages/Admin/EnterPhoneNumber/hooks/useGetCustomer.ts
@@ -1,9 +1,9 @@
 import { useQuery } from '@tanstack/react-query';
 import { getCustomer } from '../../../../api/get';
-import { PHONE_NUMBER_LENGTH } from '../../../../constants';
 import { CustomerPhoneNumberRes } from '../../../../types/api/response';
 import { CustomerPhoneNumber } from '../../../../types/domain/customer';
 import { removeHyphen } from '../../../../utils';
+import { PHONE_NUMBER_LENGTH } from '../../../../constants/magicNumber';
 
 const useGetCustomer = (phoneNumber: string) =>
   useQuery<CustomerPhoneNumberRes, Error, CustomerPhoneNumber[]>(['customer'], {
