+++ frontend/src/pages/Admin/ManageCafe/components/CafePhoneNumber/index.tsx
@@ -1,5 +1,9 @@
 import { Input } from '../../../../../components/Input';
-import { REGEX } from '../../../../../constants/regex';
+import {
+  PHONE_NUMBER_START_WITHOUT_02,
+  PHONE_NUMBER_START_WITH_02,
+} from '../../../../../constants/magicNumber';
+import REGEX from '../../../../../constants/regex';
 import { parsePhoneNumber } from '../../../../../utils';
 import { StepTitle, Wrapper } from '../../style';
 
@@ -15,7 +19,9 @@ const CafePhoneNumber = ({ phoneNumber, inputPhoneNumber }: CafePhoneNumberProps
       <Input
         id="phone-number-input"
         placeholder="전화번호를 입력해주세요"
-        maxLength={phoneNumber.startsWith('02') ? 12 : 13}
+        maxLength={
+          phoneNumber.startsWith('02') ? PHONE_NUMBER_START_WITH_02 : PHONE_NUMBER_START_WITHOUT_02
+        }
         pattern={REGEX.phoneNumber.source}
         onChange={inputPhoneNumber}
         value={parsePhoneNumber(phoneNumber)}
