+++ frontend/src/pages/Admin/ManageCafe/hooks/useGetCouponDesign.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { getCouponDesign } from '../../../../api/get';
-import { INVALID_CAFE_ID } from '../../../../constants';
+import { INVALID_CAFE_ID } from '../../../../constants/magicNumber';
 
 const useGetCouponDesign = (cafeId: number) => {
   return useQuery({
