+++ frontend/src/pages/Admin/RewardPage/index.tsx
@@ -6,7 +6,6 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import { Spacing } from '../../../style/layout/common';
 import { getReward } from '../../../api/get';
 import { patchReward } from '../../../api/patch';
-import { INVALID_CAFE_ID } from '../../../constants';
 import { useRedirectRegisterPage } from '../../../hooks/useRedirectRegisterPage';
 import {
   MutateReq,
@@ -16,6 +15,7 @@ import {
 } from '../../../types/api/request';
 import { Reward } from '../../../types/domain/reward';
 import ROUTER_PATH from '../../../constants/routerPath';
+import { INVALID_CAFE_ID } from '../../../constants/magicNumber';
 
 const RewardPage = () => {
   const cafeId = useRedirectRegisterPage();
