+++ frontend/src/pages/Customer/HistoryPage/HistoryPage.tsx
@@ -3,6 +3,12 @@ import { useNavigate } from 'react-router-dom';
 import SubHeader from '../../../components/Header/SubHeader';
 import ROUTER_PATH from '../../../constants/routerPath';
 
+export const DATE_PARSE_OPTION = {
+  hasYear: false,
+  hasMonth: true,
+  hasDay: true,
+};
+
 interface HistoryPageProps {
   title: string;
 }
