+++ frontend/src/pages/Customer/InputPhoneNumber/hooks/useGetCustomerRegisterType.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { getCustomerRegisterType } from '../../../../api/get';
-import { PHONE_NUMBER_LENGTH } from '../../../../constants';
 import { removeHyphen } from '../../../../utils';
+import { PHONE_NUMBER_LENGTH } from '../../../../constants/magicNumber';
 
 const useGetCustomerRegisterType = (phoneNumber: string) => {
   return useQuery(['customerRegisterType'], {
